+++ demo/Main.hs
@@ -66,7 +66,8 @@ loopFor time m =
 
 testCamera :: IO ()
 testCamera =
-  do initFFmpeg
+  do initFFmpeg -- Defaults to quiet (minimal) logging
+     -- setLogLevel avLogInfo -- Restore standard ffmpeg logging
      (getFrame, cleanup) <- imageReader (Camera "0:0" defaultCameraConfig)
      frame1 <- getFrame
      case frame1 of